+++ src/common/js/configure.js
@@ -1,5 +1,5 @@
 /*
- * pwix:package-template/src/common/js/configure.js
+ * pwix:admin-first/src/common/js/configure.js
  */
 
 pckTemplate.configure = function( o ){
@@ -7,6 +7,6 @@ pckTemplate.configure = function( o ){
 
     // be verbose if asked for
     if( pckTemplate._conf.verbosity & PCK_VERBOSE_CONFIGURE ){
-        console.debug( 'pwix:package-template configure() with', o, 'building', pckTemplate._conf );
+        console.debug( 'pwix:admin-first configure() with', o, 'building', pckTemplate._conf );
     }
 }