+++ src/common/js/index.js
@@ -1,5 +1,5 @@
 /*
- * pwix:package-template/src/common/js/index.js
+ * pwix:admin-first/src/common/js/index.js
  */
 
 import './pck-tempkate.js';