+++ src/common/js/constants.js
@@ -1,16 +1,16 @@
 /*
- * pwix:admin-first/src/common/js/constants.js
+ * pwix:startup-app-admin/src/common/js/constants.js
  */
 
 // exported
 
-AF_APP_ADMIN_ROLE = 'APP_ADMIN';
+SAA_APP_ADMIN_ROLE = 'APP_ADMIN';
 
-AF_VERBOSE_NONE = 0;
-AF_VERBOSE_CONFIGURE = 0x01 <<  0;
+SAA_VERBOSE_NONE = 0;
+SAA_VERBOSE_CONFIGURE = 0x01 <<  0;
 
 // non exported
 
-PACKAGE_NAME = 'pwix:admin-first';
+PACKAGE_NAME = 'pwix:startup-app-admin';
 
 I18N = PACKAGE_NAME;