+++ src/server/js/check_npms.js
@@ -1,5 +1,5 @@
 /*
- * pwix:admin-first/src/server/js/check_npms.js
+ * pwix:startup-app-admin/src/server/js/check_npms.js
  */
 
 import { checkNpmVersions } from 'meteor/tmeasday:check-npm-versions';
@@ -12,5 +12,5 @@ if( false ){
 checkNpmVersions({
     'merge': '^2.1.1'
 },
-    'pwix:admin-first'
+    'pwix:startup-app-admin'
 );