+++ test/js/index.js
@@ -1,11 +1,11 @@
 // Import Tinytest from the tinytest Meteor package.
 import { Tinytest } from "meteor/tinytest";
 
-// Import and rename a variable exported by admin-first.js.
-import { name as packageName } from "meteor/pwix:admin-first";
+// Import and rename a variable exported by startup-app-admin.js.
+import { name as packageName } from "meteor/pwix:startup-app-admin";
 
 // Write your tests here!
 // Here is an example.
-Tinytest.add( 'admin-first - example', function( test ){
-    test.equal( packageName, 'admin-first' );
+Tinytest.add( 'startup-app-admin - example', function( test ){
+    test.equal( packageName, 'startup-app-admin' );
 });