+++ package.js
@@ -1,12 +1,8 @@
 Package.describe({
     name: 'pwix:startup-app-admin',
     version: '1.0.0',
-    // Brief, one-line summary of the package.
-    summary: '',
-    // URL to the Git repository containing the source code for this package.
-    git: '',
-    // By default, Meteor will default to using README.md for documentation.
-    // To avoid submitting documentation, set this field to null.
+    summary: 'Force the application to start by creating an administrator',
+    git: 'https://github.com/trychlos/pwix-startup-app-admin',
     documentation: 'README.md'
 });