+++ src/common/js/i18n.js
@@ -9,3 +9,7 @@ pwixI18n.namespace( I18N, 'en', pwixSAA.i18n.en );
 
 import '../i18n/fr.js';
 pwixI18n.namespace( I18N, 'fr', pwixSAA.i18n.fr );
+
+pwixSAA.i18n.namespace = function(){
+    return I18N;
+};