+++ ChangeLog.md
@@ -4,10 +4,10 @@
 
 ### 1.0.0
 
-Release date: xxxxx
+Release date: 2023- 6- 9
 
 - Initial release
 
 ---
 P. Wieser
-- Last updated on 2023, May 11th
+- Last updated on 2023, June 9th